refactor: split LocalLlmTab into its two disjoint views (#5720)
The Local LLMs tab was one 1990-line file rendering two surfaces that share
nothing but the pills above them: the llama.cpp/MTPLX/Slotstream Runtimes
manager and the Hugging Face Model Library. Every state hook, effect and
handler for both was evaluated whichever view was open, and five effects
opened by bailing out on the wrong one, so a change to catalog filtering
still walked a reviewer through runtime process management.

Runtimes and Model Library now each own a file and mount only when selected.
LocalLlmTab is a 60-line dispatcher: the pills, the blurb, and which view to
render. The installed-model status both views read, and the one
busy/toast/refresh path every action routes through, moved into a shared
useLocalLlmStatus hook; because the two views never mount together, the
mounted one is still the only subscriber to the `localLlm:progress` channel.
The Ollama/LM Studio backend list both views render moved to
client/src/lib/localLlmBackends.js.

One visible fix falls out of the split: the "Upgrading Ollama" auto-upgrade
banner used to render inside the Runtimes panel, while the only thing that
can start that flow is a model install in the Model Library — so it was
unreachable. It now renders where the install that triggers it lives.
